% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_about_ui.R
\name{about_module}
\alias{about_module}
\alias{about_ui}
\alias{about_server}
\title{About Tab}
\usage{
about_ui(id)

about_server(id)
}
\arguments{
\item{id}{the shiny namespace id name (i.e., \code{"about"}).}
}
\description{
Provides the about tab within the app.
}
\details{
\code{about_ui()} a shiny UI.

\code{about_server()} a shiny module server.
}
