% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_add.R
\name{add_functions}
\alias{add_functions}
\alias{add_new_id}
\alias{add_source_id}
\alias{add_valid_cols}
\alias{add_sub_sor_tbl}
\alias{add_taxonomic_groups}
\alias{add_valid_taxonomy}
\title{Add functions}
\usage{
add_new_id(df, id_name, max_ids)

add_source_id(tbl_samples, tbl_sources)

add_valid_cols(df, valid_values)

add_sub_sor_tbl(split_tables, sub_tbl, sor_tbl)

add_taxonomic_groups(df, species_list)

add_valid_taxonomy(df, species_list)
}
\arguments{
\item{df}{a \code{data.frame}}

\item{id_name}{the id column name to add new id to in the database}

\item{max_ids}{the max id value}

\item{tbl_samples}{\code{data.frame} that contains sample data}

\item{tbl_sources}{\code{data.frame} that contains source data}

\item{valid_values}{a \code{data.frame} containing valid values from database
schema derived from \code{get_valid_values()}.}

\item{split_tables}{a \code{list} cotaining \code{tbl_samples} split into
database tables prior to submission.}

\item{sub_tbl}{the \code{data.frame} that will be submitted to \code{tbl_submission}
in the database.}

\item{sor_tbl}{the \code{data.frame} that will be submitted to \code{tbl_sources}
in the database.}

\item{species_list}{a \code{tbl_lazy} object of the \code{tbl_taxonomy} from the
database}
}
\description{
These functions add new columns to the ingested
data prior to being validated or imported
into the database
}
\details{
\code{add_valid_taxonomy()} adds columns whether or not \code{valid_taxonomy()}
returns species that are present in \code{tbl_taxononomy}'s \code{common_name} and
\code{scientific_name}
}
\seealso{
\code{\link[=get_valid_values]{get_valid_values()}}
}
