% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_app_version.R
\name{App version}
\alias{App version}
\alias{app_version_head}
\alias{app_version_label}
\title{App version functions}
\usage{
app_version_head()

app_version_label(app_version, build_time = NULL)
}
\description{
This function, which has no arguments, dynamically displays
the app version and build time.
}
