% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_check.R
\name{check_functions}
\alias{check_functions}
\alias{check_empty_character}
\alias{check_sheets}
\alias{check_email}
\alias{check_tab_credentials}
\alias{check_taxonomy_match}
\title{Check functions}
\usage{
check_empty_character(x)

check_sheets(file_path, output)

check_email(x, arg_name = NULL)

check_tab_credentials(user, valid_users)

check_taxonomy_match(input_values, db_values)
}
\arguments{
\item{x}{a `vector``}

\item{file_path}{the file path supplied by reactive handler}

\item{output}{shiny object to output too}

\item{arg_name}{the name of the argument}

\item{user}{a \code{vector} containing the username}

\item{valid_users}{a \code{vector} containinng valid user}

\item{input_values}{a \code{vector} of common and species names in incoming
excel sheet.}

\item{db_values}{a \code{vector} of common and species naems that are
in \code{tbl_taxonomy}}
}
\description{
These functions check different conditions
}
