% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_clean.R
\name{clean_functions}
\alias{clean_functions}
\alias{clean_all_validations}
\alias{clean_data_tables}
\alias{clean_db_constraints}
\alias{clean_match_to_db_col}
\alias{clean_rename_to_db_col}
\title{Clean functions}
\usage{
clean_all_validations(...)

clean_data_tables(df, flag_cols, type, group_cols, filter_coords = TRUE)

clean_db_constraints(clause)

clean_match_to_db_col(col_name, db_cols)

clean_rename_to_db_col(df, con, table_name)
}
\arguments{
\item{...}{objects to be cleaned. Usually validation reports}

\item{df}{a \code{data.frame} of imported new data}

\item{flag_cols}{collumns to flag and coalesce}

\item{type}{is the data table type that is being supplied e.g., tbl_isotope}

\item{group_cols}{columns that will be grouping our summariziing e.g.,
lat and lon, pi name, waterbody ect.}

\item{clause}{constraint clause to be cleaned}

\item{col_name}{column name to be cleaned to match database column names}

\item{db_cols}{column names in the database to be used}

\item{con}{PostgreSQL connection object using \code{{DBI}}}

\item{table_name}{the table name in the database that each
column name is assinged to.}
}
\description{
These functions clean up \code{data.frames} and \code{vectors} that
are coming into the the databae prior to submission,
validation reports, or are in the database and need need to be
cleaned prior to being displayed (e.g., schema constraints)
}
