% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_clear.R
\name{clear_functions}
\alias{clear_functions}
\alias{clear_fields}
\alias{clear_table}
\title{Clear functions}
\usage{
clear_fields(session, cols)

clear_table(cols)
}
\arguments{
\item{session}{session to be cleared}

\item{cols}{column names}
}
\description{
These functions allow fields to be cleared within a server session.
}
