% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_convert.R
\name{convert_functions}
\alias{convert_functions}
\alias{convert_nice_name}
\title{Convert functions}
\usage{
convert_nice_name(cols, lookup = nice_name_lookup)
}
\arguments{
\item{cols}{column to be converted}

\item{lookup}{a \code{vector} that has the name of the column
and the nice looking name of the column e.g., \code{sample_id}
becomes \verb{Sample ID}.}
}
\description{
These functions convert columns whether that is column names or
the values in the columns themeselves.
}
