% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_create.R
\name{create_functions}
\alias{create_functions}
\alias{create_filtered_data}
\alias{create_mean_data}
\alias{create_raw_data}
\alias{create_searching_data}
\title{Create functions}
\usage{
create_filtered_data(data, input_source, pane)

create_mean_data(data, input_source)

create_raw_data(
  activated = NULL,
  con,
  input_source,
  main_input,
  tab = NULL,
  var_field
)

create_searching_data(
  collect = TRUE,
  con = NULL,
  input,
  source_data = NULL,
  tbl_name = NULL
)
}
\arguments{
\item{data}{a \code{reactive()} data frame like object usually
created from another \code{create_function}. Considering
raw data is gathered through a PostgresSQL connection
these reactive objects tend to be \code{tbl_lazy}.}

\item{input_source}{usually an object created by a sidebar
function. These objects tend to be \code{reactive()} outcomes from
\code{observe()} or \code{observeEvent()} calls within a module.}

\item{pane}{the sidepane that is bein selected e.g.,
\verb{"scatter_plot". }}

\item{activated}{is a \code{shiny} reactive value which gets triggered
when the tab is activated. This makes it so that the tab stays the
same when clicking away but also that it doesn't run
the tab upon start up nor when clicking away.}

\item{con}{a \code{DBI} conection to, in this case PostgreSQL database}

\item{main_input}{the \code{main_input} from the shiny server.}

\item{tab}{the selected tab that is being displayed e.g.,
\verb{"scatter_plot". }}

\item{var_field}{a \code{character} string that identifies the object
in the \code{list} that is supplied to \code{input_source}. This object
name is the variable field that is of interest for the particular
use for a given function.}

\item{collect}{a \code{logical} that determines whether to \code{collect()}
the data from the PostgreSQL database and turn into a \code{tibble}
object in a given session. Default is \code{TRUE}.`}

\item{input}{a given shiny module and server input.}

\item{source_data}{a reactive object that is \code{tbl_source}
a \code{tbl_lazy} object from a PostgreSQL database.}

\item{tbl_name}{a \code{character} string of the database name
to create searching data from e.g., \code{tbl_taxonomy}.}
}
\description{
These functions are all use other functions to create
\code{shiny::reactive()} objects usually used in modules.
Any function that uses a \code{reactive()} call is considered
to be creating an object that is then displayed in the UI.
}
