% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_start_db_connection.R
\name{db_connections}
\alias{db_connections}
\alias{start_db_con}
\title{Start Database Connection}
\usage{
start_db_con(username = NULL)
}
\arguments{
\item{username}{The username for user specific connections.
Default is \code{NULL} and will use default user.}
}
\value{
returns a connection object.
}
\description{
This function uses environment supplied variables, \code{{DBI}},
and \code{{RPostgres}} to connect to a PostgreSQL database.
}
