% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_display.R
\name{display_functions}
\alias{display_functions}
\alias{display_add_taxa}
\alias{display_col_width}
\alias{display_hist}
\alias{display_scatter_plot}
\alias{display_submission_map}
\alias{display_sub_map_msg}
\alias{display_table}
\alias{display_upload_status}
\alias{display_validation_status}
\title{Display functions}
\usage{
display_add_taxa(data, output, output_id, session)

display_col_width(...)

display_hist(data, input_source, output, output_id = "summary_histogram")

display_scatter_plot(data, input_source, output, output_id = "scatter_plot")

display_submission_map(ns, output, output_id = "map", split_tables)

display_sub_map_msg(
  ns,
  output,
  output_id = "location_map",
  split_tables,
  validated_submission,
  validated_sources,
  validated_samples
)

display_table(data, output, output_id = "summary_table_output", search = TRUE)

display_upload_status(
  ns,
  output,
  output_id = "upload_status",
  upload_succeeded = NULL,
  submission_results = NULL,
  email_succeeded = NULL,
  user_email = NULL
)

display_validation_status(
  ns,
  output,
  output_id = "upload_status",
  split_tables = NULL,
  validated = TRUE
)
}
\arguments{
\item{data}{a reactive data object for a given variable.
\code{observe()} or \code{observeEvent()} calls within a module.}

\item{output}{an output within a shiny module.}

\item{output_id}{the id name of the output, e.g., \code{"summary_histogram"}.}

\item{session}{module session.}

\item{...}{additional arguments}

\item{input_source}{usually an object created by a sidebar
function. These objects tend to be \code{reactive()} outcomes from
\code{observe()} or \code{observeEvent()} calls within a module.}

\item{ns}{a namespace object created from \code{NS()} from \code{shiny}}

\item{split_tables}{a \code{list} that contains the tables that will
be submitted split into their given PostgreSQL table names.}

\item{validated_submission}{a reactive value this contains the
validated \code{tbl_submission}.}

\item{validated_sources}{a reactive value this contains the
validated \code{tbl_sources}.}

\item{validated_samples}{a reactive value this contains the
validated \code{tbl_samples}.}

\item{search}{a logitcal value that determines whether \code{datatable()} from
\code{DT} has a search bar or not. Default value is \code{TRUE}.}

\item{upload_succeeded}{a \code{tryCatch()} that tries to submit the split tables
to the database. The returning object is a logical value, \code{TRUE} or \code{FALSE}
which will change the given status of ghe upload that is then displayed to the user.
This object is returned by \code{upload_to_db()}.}

\item{submission_results}{an object returned by \code{upload_to_db} that has the
name of the table, the number of rows that were submitted, and
the submission id so this can be dynamically displayed to the user.}

\item{email_succeeded}{a \code{tryCatch()} that tries to send an meail. The returning
object is a logical value, \code{TRUE} or \code{FALSE} which will change the given email
status displayed to the user.}

\item{user_email}{a valid email address of the user, e.g., \code{user.name@example.com}.}

\item{validated}{a logical value to determine if a validation
succeeded. Default is \code{TRUE}.}
}
\description{
These functions are all display vital information
within GLATAR such as plots, histograms, and tables.
They function by wrapping the given display object (e.g.,
table) with a \verb{render*()} function from \code{shiny}.
}
