% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_observer.R
\name{exclusive_all_observer}
\alias{exclusive_all_observer}
\title{Exclude all observer}
\usage{
exclusive_all_observer(input, session, id)
}
\arguments{
\item{input}{a server input value.}

\item{session}{a shiny session.}

\item{id}{the dropdown object to apply the observer}
}
\value{
an Event that excludes `"All"``
}
\description{
This function uses an \code{observerEvent()} to adjust dropdown
menus so that when other selections beside \code{"All"} is selected
it will not allow \code{"All"} to be selected until the entire box
is cleared.
}
