% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_export.R
\name{export_functions}
\alias{export_functions}
\alias{export_excel}
\title{Exporf functions}
\usage{
export_excel(
  input_source,
  file_name,
  output,
  output_id,
  tabs,
  id,
  input,
  session,
  df_name
)
}
\arguments{
\item{input_source}{is the shiny input source}

\item{file_name}{file name to be used}

\item{output}{a shiny output}

\item{output_id}{a shiny output id e.g., \code{download_summary}}

\item{tabs}{tab name to be used e.g., \code{summary_info}}

\item{id}{shiny id object}

\item{input}{a shiny input object}

\item{session}{a shiny session object}

\item{df_name}{the name of the df to be exported}
}
\description{
These functions allow the exportation of different objects usually
into excel files.
}
