% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_fix.R
\name{fix_functions}
\alias{fix_functions}
\alias{fix_case_types}
\alias{fix_logical_types}
\alias{fix_table_order}
\alias{fix_title_label}
\alias{fix_var_generic}
\title{Fix Functions}
\usage{
fix_case_types(df)

fix_logical_types(x, target_type)

fix_table_order(split_tables)

fix_title_label(x, max = NULL)

fix_var_generic(df, var_raw)
}
\arguments{
\item{df}{a \code{data.frame}}

\item{x}{a vector of selected variables.}

\item{target_type}{a \code{character} indcaiting the type to convert into e.g. \code{"as.numeric"}.}

\item{split_tables}{a \code{list} containing inported data as \code{data.frames} ready
to be submited. `}

\item{max}{the maximum nmber of objects to permit
before the function starts condense the names. Defaults to \code{4}.}

\item{var_raw}{a reactive value that is the raw variable name}
}
\description{
These functions fix or adjust an object to better fit within
the required user interface and database. For example case types
}
\details{
\code{fix_case_types} fix case types once validation has occured for incoming
data into the database.

\code{fix_logical_types} fix logic types that are blank and \code{NA}

\code{fix_table_order} fixes the order of the tables within the \code{list} so we
can iteratively load data into the the database. The proper order is
\code{tbl_submission}, \code{tbl_sources}, and `tbl_samples`` the an other supporting
tables.

\code{fix_title_label} adjusts the plot title to be dynamic and change
based on the selected species and waterbodies.

\code{fix_var_generic} returns a \code{list} containing three objects one being \code{df} with
the selected variable, one being the \code{variable} name and the last being
\verb{variable label}. This function is used in summary tables, histograms and scatter plots.
}
