% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_get.R
\name{get_functions}
\alias{get_functions}
\alias{get_column_map}
\alias{get_data}
\alias{get_data_tables}
\alias{get_dropdown_choices}
\alias{get_good_groups}
\alias{get_id_col}
\alias{get_id_max}
\alias{get_join_table}
\alias{get_numeric_vars}
\alias{get_ranges}
\alias{get_raw_data}
\alias{get_selected_tab}
\alias{get_summary_data}
\alias{get_submission_id}
\alias{get_tables_needed}
\alias{get_taxa_col}
\alias{get_theme_choices}
\alias{get_valid_values}
\alias{get_var_types}
\title{Get functions}
\usage{
get_column_map(con)

get_data(con, debug_sql = FALSE)

get_data_tables(con, df, data_tables, flag_cols, var)

get_dropdown_choices(df, type)

get_good_groups(df)

get_id_col(con)

get_id_max(table_name, id_col)

get_join_table(df, table, con)

get_numeric_vars(con)

get_ranges(df, x_var, y_var)

get_raw_data(
  con,
  selected_vars = NULL,
  grouping_vars = NULL,
  debug_sql = FALSE
)

get_selected_tab(input)

get_summary_data(
  con,
  selected_vars = NULL,
  grouping_vars = NULL,
  debug_sql = FALSE
)

get_submission_id(con)

get_tables_needed(con, var)

get_taxa_col(con)

get_theme_choices(
  con,
  theme,
  numeric_choices,
  numeric_names,
  length_vars,
  energy_vars
)

get_valid_values(con)

get_var_types(df, var)
}
\arguments{
\item{con}{a valid \code{DBI} connection to a PostgreSQL database.}

\item{debug_sql}{a logical value that will provide SQL for
debugging. Defualt is \code{FALSE}.}

\item{df}{a \code{tbl_lazy}.}

\item{data_tables}{a vector produced by \code{data_tables()}.}

\item{flag_cols}{a vector that has replaced \code{tbl_} to \code{has_}
if table is present in \code{data_tables()}.}

\item{var}{a variable of interst to get unique values.}

\item{type}{variable of interest for a given dropdown.}

\item{table_name}{the name of the database table to get the maxiumum
id column value.}

\item{id_col}{name of the corresponding id column in a given \code{table_name}
that we want the maximum id value.}

\item{table}{table name to get from the database.}

\item{x_var}{the \code{x} variable of interest.}

\item{y_var}{the \code{y} variable of interest.}

\item{selected_vars}{a \code{reactive} object that is the column name of interest
usually generated from sidebar.}

\item{grouping_vars}{a \code{reactive} object that is the column name of grouping
variables usually generated from sidebar.}

\item{input}{input from UI}

\item{theme}{a \code{vector} containing different data themes}

\item{numeric_choices}{a \code{reactive} value that has numerical choices.}

\item{numeric_names}{a \code{reactive} value that has selected numerical column names.}

\item{length_vars}{a \code{reactive} value that has length variables.}

\item{energy_vars}{a \code{reactive} value that has energy variables.}
}
\description{
These functions are very important as they get infomration
from a PostgreSQL database, whther that is
schemas, raw values or other information about the database.
Each function creates a SQL string that is excuted on a
given table in the database.
}
\details{
\code{get_column_map()} gets the data dictonary from the database.

\code{get_data()} retrives base template for all sample data in the
the database. It gets \code{tbl_samples}, \code{tbl_loc}, and \code{tbl_length}
and joins them to make a base for all other opperations.

\code{get_data_tables()} this retrieves and identifes which data belongs to each sample
id. For example, id \code{1} has calorimetry and proximent composition data. This will
start the base dataframe to gather this information.

\code{get_id_col()}use a SQL query where we select all of the ID columns
throughout the entire database and then select only the columns that
we want to add new ID values to. This information is then fed to
\code{get_id_max()}.

\code{get_raw_data()} retrives base template for all sample data in the
the database and joins variables and grouping variables of interest.

\code{get_selected_tabs()} is an exception within \verb{get_*} functions as
it doesn't interact with the database and instead interacts with
the \code{shiny} UI.

\code{get_summary_data()} retrives base template for all sample data in the
the database and joins variables and grouping variables of interest for
summarizing.

\code{get_tables_needed()} this gets the tables of interest

\code{get_valid_values()} gets all the schemas in the database and
grabs all the constraints and cleans them up so we can use the
constraints to validate data.
}
\seealso{
\code{\link[=get_id_max]{get_id_max()}}

\code{\link[=get_id_col]{get_id_col()}}

\code{\link[=get_column_map]{get_column_map()}}
}
