% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{glatar_app}
\alias{glatar_app}
\title{GLATAR Shiny App}
\usage{
glatar_app()
}
\description{
This function which has no arguments starts up the app,
creates the user interface (UI) and server and combines them
to make the app.
}
