% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_glatar_diagram.R
\name{glatar_diagram}
\alias{glatar_diagram}
\alias{glatar_diagram_ui}
\title{GLATAR diagram}
\usage{
glatar_diagram_ui()
}
\description{
The \code{glatar_diagram_ui()} create the interactive
HTML solar system on the front page.
}
