% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_how_to_use_ui.R
\name{how_to_server_module}
\alias{how_to_server_module}
\alias{how_to_server}
\title{How to Use Server}
\usage{
how_to_server(id, parent_session)
}
\arguments{
\item{id}{the shiny namespace id name (i.e., \code{"how_to_use"}).}

\item{parent_session}{a shiny server session.}
}
\value{
\code{how_to_server()} a shiny module server.
}
\description{
Provides the server for the documentation tab
}
