% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_load_indicator.R
\name{load_indicator}
\alias{load_indicator}
\alias{load_indicator_hide}
\title{Load Indicator functions}
\usage{
load_indicator(input, output)

load_indicator_hide(input, output)
}
\arguments{
\item{input}{a shiny \code{server} input value.}

\item{output}{a shiny \code{server} output value.}
}
\description{
Load Indicator functions
}
\details{
\code{load_indicator()} displays load indicator during file uploading

\code{load_indicator_hide()} hides load indicator.
}
