% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_login_ui.R
\name{login_modal_server}
\alias{login_modal_server}
\alias{tab_auth_server}
\alias{logout_server}
\title{Authorization Server}
\usage{
tab_auth_server(
  input,
  output,
  session,
  valid_users_emails,
  sidebar_id = "tabs",
  off = FALSE
)

logout_server(id, parent_session)
}
\arguments{
\item{input}{a shiny input object.}

\item{output}{a shiny output object.}

\item{session}{a shiny session.}

\item{valid_users_emails}{a \code{vector} containing the valid user email}

\item{sidebar_id}{a \code{vector} containing the sidebar ids to allow
the login to display.}

\item{off}{a \code{logical} value that turns on and off the authorization
server. Default is \code{FALSE}. This switch is for development only.}

\item{id}{the shiny namespace id name (i.e., \code{"logout"}).}

\item{parent_session}{a shiny server session.}
}
\description{
Provides the authorization server for the login modal that displays
on two locked tabs, uploading and user specific data.
}
\details{
\code{tab_auth_server()} a shiny modal

\code{logout_server()} a modal that asks the user if they really want to logout.
This prevents accidental logouts.
}
