% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_make.R
\name{make_functions}
\alias{make_functions}
\alias{make_pg_username}
\alias{make_temp_password}
\title{Make functions}
\usage{
make_pg_username(email)

make_temp_password(email)
}
\arguments{
\item{email}{a valid email}
}
\description{
These functions make information that is to be used during the
login processes
}
\details{
\code{make_pg_username()} returns a username that matches with the
database user name.

\code{make_temp_password()} returns a temporay password
}
