% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_msg.R
\name{msg_functions}
\alias{msg_functions}
\alias{msg_dropdowns}
\alias{msg_hist_ui}
\alias{msg_hist_vars}
\alias{msg_input_source}
\alias{msg_log_agent}
\alias{msg_mean_data}
\alias{msg_summary_data}
\alias{msg_selected_vars}
\alias{msg_tab_name}
\title{Message functions}
\usage{
msg_dropdowns(
  waterbody_choices,
  common_name_choices,
  grouping_choices,
  numeric_choices
)

msg_hist_ui(df, var, type_val, col)

msg_hist_vars(df, var, ba)

msg_input_source(input_source_name, envir = parent.frame())

msg_log_agent(x, name, show = 10)

msg_mean_data(df, summary_grouping_vars, y_vals)

msg_summary_data(df, name = deparse(substitute(df)))

msg_selected_vars(selected_vars)

msg_tab_name(tab)
}
\arguments{
\item{waterbody_choices}{a vector of water body choices}

\item{common_name_choices}{a vector of common name choices}

\item{grouping_choices}{a vector of grouping variable choices}

\item{numeric_choices}{a vector of numerical choices}

\item{df}{the \code{data.frame} that has summary information.}

\item{var}{the \code{x} variable that is being supplied to the histogram}

\item{type_val}{the type of value supplied}

\item{col}{the column name of the column in
dataframe that is of interest}

\item{ba}{vector that is either \code{"before"} or \code{"after"}.}

\item{input_source_name}{the name of the input}

\item{envir}{the R environment}

\item{x}{an \code{validator} object}

\item{name}{the object name that has been triggered. It
is usually a \code{reactive} object.}

\item{show}{how many values to display, default is \code{10}}

\item{summary_grouping_vars}{the summary grouping variables}

\item{y_vals}{the \code{y} value of interst}

\item{selected_vars}{an object that has selected variables from
dropdowns.}

\item{tab}{the name of the tab.}
}
\description{
These functions use \code{{cli}} to display messages in other functions
when different opperations occur. They are helpfui in seeing
how the suer interacts with the app as well as diagnosing
isssues that may arrise.
}
\details{
\code{msg_dropdowns()} returns info on the drop downs selected.

\code{msg_hist_ui()} provides info on the histogogram

\code{msg_hist_vars()} provides info on the histogogram variables

\code{msg_input_source()} returns info on whether info from valid sources is being used.

\code{msg_log_agent()} returns info on validator agents
and whether they were successful.

\code{msg_mean_data()} returns info on the mean data object

\code{msg_summary_data()}returns info about summary data

\code{msg_selected_vars()}returns info about the selected varialbes

\code{msg_tab_name()}tab name.
}
