% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naming_conventions}
\alias{naming_conventions}
\title{A \code{data.frame} containing naming conventions}
\format{
\code{data.frame} containing 87 rows and 2 variables
\describe{
\item{raw_names}{name database columns}
\item{nice_names}{nice names to transform them into}
}
}
\usage{
naming_conventions
}
\description{
A \code{data.frame} containing naming conventions
}
\keyword{datasets}
