% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_plot_control.R
\name{plot_controls}
\alias{plot_controls}
\alias{plot_ui}
\alias{pallete_selector}
\alias{pallete_selector_server}
\alias{alpha_selector}
\alias{alpha_selector_server}
\alias{size_selector}
\alias{size_selector_server}
\alias{shape_selector}
\alias{shape_selector_server}
\alias{zoom_slider_ui}
\alias{zoom_slider_server}
\title{Plot control functions}
\usage{
plot_ui(title, plot_id, height, ...)

pallete_selector(...)

pallete_selector_server(input)

alpha_selector(...)

alpha_selector_server(input)

size_selector(...)

size_selector_server(input)

shape_selector(...)

shape_selector_server(input)

zoom_slider_ui(ns, id, label)

zoom_slider_server(input)
}
\arguments{
\item{title}{the title of plot}

\item{plot_id}{the shiny namespace for the plot desired \code{"scatter_plot"}}

\item{height}{the height of the plot in pixels}

\item{...}{additional parameters to be given for example a \code{NS()} object.}

\item{input}{a shiny server input value.}

\item{ns}{a shiny namespace object e.g., \code{ns}.}

\item{id}{a shiny namespace object e.g., \code{"zoom_x"}.}

\item{label}{a shiny namespace object e.g., \code{"Zoom X Axis"}.}
}
\description{
These functions add different control features to different
plots displayed such as scatter and eventually box plot.
}
\details{
\code{plot_ui()} provides user interface for any plot object throughout
the app.

\code{pallete_selector()} provides user interface to select different palletes
from \href{https://CRAN.R-project.org/package=viridis}{viridis}.

\code{pallete_selector_server()} provides the server and reactive value
to select different palletes from \href{https://CRAN.R-project.org/package=viridis}{viridis}.

\code{alpha_selector()} provides user interface to select different values
of \code{alpha} which controls transparency.

\code{alpha_selector_server()} provides the server and reactive value to
select different values of \code{alpha} which controls transparency.

\code{size_selector()} provides user interface to select different values
of \code{size} which controls the size of the points.

\code{size_selector_server()} provides the server and reactive values
of \code{size} which controls the size of the points.

\code{shape_selector()} provides user interface to select different values
of \code{shapes} which controls the shape type.

\code{shape_selector_server()} provides the server and reactive values
of \code{size} which controls the size of the points.

\code{zoom_slider_ui()} provides the the user interface for
zoom slider on the plot panel.

\code{zoom_slider_server()} provides the server and reactive values
of zoom slider which controls the zoom on the x and y axis.
}
