% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_plot_message.R
\name{plot_message}
\alias{plot_message}
\title{Plot message functions}
\usage{
plot_message(msg)
}
\arguments{
\item{msg}{A character vector containing the message that is desired.}
}
\value{
a \code{ggplot()} object to be displayed.
}
\description{
This function allows for \href{https://ggplot2.tidyverse.org/}{ggplot2} to be used to display blank plots
with messages to the user.
}
