% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_ram_tracker.R
\name{ram_tracker}
\alias{ram_tracker}
\title{Ram Tracker}
\usage{
ram_tracker()
}
\value{
displays the amont of RAM used during
app deployment. Checks RAM usage every 5 seeconds
and returns values in mb.
}
\description{
Tracks the amoutn of RAM used during shiny app
deployment.
}
