% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_raw_sidebar.R
\name{raw_data_module}
\alias{raw_data_module}
\alias{raw_data_sidebar_ui}
\alias{raw_data_sidebar_server}
\title{Raw Data Sidebar}
\usage{
raw_data_sidebar_ui(id)

raw_data_sidebar_server(id, con, main_input, auth_state)
}
\arguments{
\item{id}{the shiny namespace id name (i.e., \code{"raw_sidebar"}).}
}
\value{
\code{raw_data_sidebar_ui()} that provides user interface for raw data
sidebar

\code{raw_data_sidebar_ui()} that provides user interface for raw data
sidebar
}
\description{
Provides the sidebar for the raw data pane.
}
