% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_read.R
\name{read_functions}
\alias{read_functions}
\alias{read_col_types}
\alias{read_xl}
\title{Read functions}
\usage{
read_col_types(file_path, tbl_name, skip, n_max)

read_xl(
  file_path,
  tbl_name,
  con = NULL,
  col_types = NULL,
  skip,
  rename = TRUE,
  rename_twice = NULL
)
}
\arguments{
\item{file_path}{the path to the file provided by UI upload observer.}

\item{tbl_name}{the table name to be imported}

\item{skip}{the number of rows to skip given the header is not the first row.}

\item{n_max}{n_max is number of the maximum number of data rows to read.}

\item{con}{a \code{DBI} conection to, in this case PostgreSQL database.}

\item{col_types}{a vector contianing the column types supplied by \code{read_col_types()}.}

\item{rename}{a logical that strips excesive information and renames column names
to snake case and follows the database shcema. Defaults to \code{TRUE}.}

\item{rename_twice}{a logical that strips excesive information and renames column names
to snake case and follows the database shcema. Defaults to  `NULL``.}
}
\value{
\code{read_col_types()} returns a vector that has each column type
for a given excel file.

\code{read_xl()}returns a \code{data.frame} with the a given database table.
}
\description{
Read functions allow for \code{.xlsx} files to be read into the the app
to be able to then manipulate and eventually validate and import into
the database.
}
\details{
`read_col_types()`` reads in a file and determiens the number of
columns and the column types

`read_xl()`` reads in a file and cleans up the columns names before any other
processes occure such as validation and spliting to the database.
}
