% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_email.R
\name{reg_confirmation_email_body}
\alias{reg_confirmation_email_body}
\title{Registration Confirmation Email Body}
\usage{
reg_confirmation_email_body(first, last, affil, email)
}
\arguments{
\item{first}{a \code{vector} containing the first name of the submitter.}

\item{last}{a \code{vector} containing the last name of the submitter.}

\item{affil}{a \code{vector} containing the affiliation of the submitter.}

\item{email}{a \code{vector} containing the email of the submitter.}
}
\value{
HTML of the email structure to be sent for the registration email.
}
\description{
This function drafts the text body the registration confirmation
email body.
}
