% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_register_feilds.R
\name{register_fields}
\alias{register_fields}
\title{Registration Feilds}
\usage{
register_fields(...)
}
\arguments{
\item{...}{additional parameters to be given for example a \code{NS()} object.}
}
\value{
shiny UI objects
}
\description{
This function produces the UI for the registraiton fields both
in the modal and tab
}
