% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_observer.R
\name{register_observer}
\alias{register_observer}
\title{Registration Observer}
\usage{
register_observer(input, output, session, submission_flag)
}
\arguments{
\item{input}{a server input value.}

\item{output}{a server output value.}

\item{session}{a shiny session.}
}
\value{
an event that sends an email with registeration information.
}
\description{
This function uses an \code{observerEvent()} to provide
server side events either in the login modal or
the registration tab.
}
