% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_register.R
\name{register_server}
\alias{register_server}
\title{Registration Server}
\usage{
register_server(id)
}
\arguments{
\item{id}{the shiny namespace id in this case
it is `"register"``.}
}
\value{
a shiny server object that contains
registration server.
}
\description{
Porvides the server side for the
\code{register_ui()}.
}
