% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_register.R
\name{register_ui}
\alias{register_ui}
\title{Registration User Interface}
\usage{
register_ui(id)
}
\arguments{
\item{id}{the shiny namespace id in this case
it is `"register".}
}
\value{
a shiny dashboard interface
}
\description{
This function provides the user interface
to be able to register to GLATAR.
}
