% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_validation_rules.R
\name{rule_functions}
\alias{rule_functions}
\alias{rule_blank}
\alias{rule_column_names}
\alias{rule_email}
\alias{rule_email_registered}
\alias{rule_len}
\alias{rule_match}
\alias{rule_na}
\title{Valdiation rules}
\usage{
rule_blank(required_fields)

rule_column_names(required_fields)

rule_email(submission_email)

rule_email_registered(submission_email)

rule_len(required_fields)

rule_match(exprs, field)

rule_na(required_fields)
}
\arguments{
\item{required_fields}{is a vector containing the fields to apply the validation
rule to.}

\item{submission_email}{the email of the submitter.}

\item{exprs}{is the validation rule expression to be checked}

\item{field}{is the column of interest, usually this is \code{col_name}
or \code{issue}/.}
}
\value{
A rule to be evaluated by \code{{validator}}.
}
\description{
These functions supply different validation rules to be used with
\code{{validator}}. They include things like a rule regarind column names,
email styles, or whether the field can be blank.
}
\details{
\code{rule_blank()} checks whether a feild is blank.

\code{rule_column_names()} checks the column names.

\code{rule_email()} checks if the submission email is in a valid format.

\code{rule_email()} checks if the submission email is in a valid format.

\code{rule_len()} checks if the length of the column equals 1.

\code{rule_match()} checks if the valid exprssion matches a given rule.

\code{rule_na()} checks if a column is \code{NA}.
}
