% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rule_map}
\alias{rule_map}
\title{A \code{data.frame} containing validation rules and names}
\format{
\code{data.frame} containing 23 rows and 3 variables
\describe{
\item{pat}{validation pattern}
\item{col}{column name that is being validated}
\item{issue}{what is the issue that the column has}
}
}
\usage{
rule_map
}
\description{
A \code{data.frame} containing validation rules and names
}
\keyword{datasets}
