% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scatter_sidebar.R
\name{scatter_plot_module}
\alias{scatter_plot_module}
\alias{scatter_sidebar_ui}
\alias{scatter_sidebar_server}
\title{Scatter Plot Sidebar}
\usage{
scatter_sidebar_ui(id)

scatter_sidebar_server(id, con, main_input)
}
\arguments{
\item{id}{the shiny namespace id name (i.e., \code{"scatter_sidebar"}).}

\item{con}{a \code{DBI} conection to, in this case PostgreSQL database.}

\item{main_input}{the shiny input from the main server}
}
\description{
Provides the sidebar for the scatter plot pane.
}
\details{
\code{scatter_sidebar_ui()} that provides user interface for scatter plot
sidebar.

\code{scatter_sidebar_server()} provides the scatter plot sidebar server.
}
