% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_search.R
\name{search_functions}
\alias{search_functions}
\alias{search_table}
\title{Search functions}
\usage{
search_table(df, search_term, search_cols)
}
\arguments{
\item{df}{the data object being supplied can either be a \code{tbl_lazy} or \code{data.frame}.}

\item{search_term}{a dynamic string that contains search terms}

\item{search_cols}{a dynamic \code{vector} containing the names of the search
columns}
}
\value{
returns a filtered data object that can then be displayed.
}
\description{
Search functions allow for a dynamic search bar to be displayed to search
taxa and source material
}
