% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_email.R
\name{send_email}
\alias{send_email}
\title{Send Email}
\usage{
send_email(to_user, email_text, attachment_path = NULL)
}
\arguments{
\item{to_user}{a valid email address}

\item{email_text}{an object containing HTML for the subject
line and the body of the email.}

\item{attachment_path}{path to attachement}
}
\description{
This function sends an email using \code{httr} using the
starndard HTTPS port of 443. Due to server firewall restrictions
we cannot use STMP protocols to send emails from the site but
can use HTTPS and resend.
}
