% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_source_sidebar.R
\name{source_sidebar_module}
\alias{source_sidebar_module}
\alias{source_sidebar_ui}
\alias{source_sidebar_server}
\title{Source Table Sidebar}
\usage{
source_sidebar_ui(id)

source_sidebar_server(id, main_input, source_sidebar_vals)
}
\arguments{
\item{id}{the shiny namespace id name (i.e., \code{"source_sidebar"}).}

\item{main_input}{the shiny input from the main server}
}
\description{
Provides the sidebar for the source table pane.
}
\details{
\code{source_sidebar_ui()} provides the summary table sidebar user interface.

\code{source_sidebar_server()} provides the summary table sidebar server.
}
