% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_split_tables.R
\name{split_tables}
\alias{split_tables}
\title{Split Tables}
\usage{
split_tables(df, tables_to_split)
}
\arguments{
\item{df}{the \code{data.frame} containing the submitted data.}

\item{tables_to_split}{the information from data dictionary
determing which columns belong to which tables.}
}
\value{
returns a \code{list}containing the tables named
apporipately and split into their proper tables for
database submission.
}
\description{
Split tables takes validated and cleaned data submissions and splits
them by columns into the table name from the data dictionary.
}
