% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_startup.R
\name{start_up}
\alias{start_up}
\title{Start up GLATAR App}
\usage{
start_up()
}
\description{
This function, which has no arguments, starts the database
connection, gets \verb{valid_values`` for validation, gets }nice_names` for properly displaying tables, creates
credientials and location for resources
}
