% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_email.R
\name{submission_email_body}
\alias{submission_email_body}
\title{Submission Email Body}
\usage{
submission_email_body(submission_id, submission_results)
}
\arguments{
\item{submission_id}{a \code{vector} containing the submission id
which will be used in the subject line and body of the emaul}

\item{submission_results}{an object returned by \code{upload_to_db}
that has the name of the table, the number of rows that were submitted, and
the submission id so this can be dynamically added to the the email text.}
}
\value{
HTML of the email structure to be sent.
}
\description{
This function drafts the text body the submission email that
will be sent once the data has been successfully uploaded to
the database
}
