% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_view_summer_info.R
\name{summary_info_module}
\alias{summary_info_module}
\alias{view_summary_info_ui}
\alias{summary_info_server}
\title{Summary Info Tab}
\usage{
view_summary_info_ui(id)

summary_info_server(id, con, main_input, summary_sidebar_vals)
}
\arguments{
\item{id}{the shiny namespace id name (i.e., \code{"summary_info"}).}

\item{con}{a \code{DBI} conection to, in this case PostgreSQL database.}

\item{main_input}{the shiny input from the main server.}

\item{summary_sidebar_vals}{a \code{reactiveVal()} object produced by \code{summary_sidebar_server()}.
The inputs from this dictate what is displayed in the main panel.}
}
\description{
Provides the summary info tab.
}
\details{
\code{view_summary_info_ui()} provides the summary info user interface.

\code{summary_info_server()} provides the summary info server.
}
