% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_summary_sidebar.R
\name{summary_sidebar_module}
\alias{summary_sidebar_module}
\alias{summary_sidebar_ui}
\alias{summary_sidebar_server}
\title{Summary Table Sidebar}
\usage{
summary_sidebar_ui(id)

summary_sidebar_server(id, con, main_input)
}
\arguments{
\item{id}{the shiny namespace id name (i.e., \code{"summary_sidebar"}).}

\item{con}{a \code{DBI} conection to, in this case PostgreSQL database.}

\item{main_input}{the shiny input from the main server}
}
\description{
Provides the sidebar for the summary table pane.
}
\details{
\code{summary_sidebar_ui()} provides the summary table sidebar user interface.

\code{summary_sidebar_server()} provides the summary table sidebar server.
}
