% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_login.R
\name{tab_login_modal}
\alias{tab_login_modal}
\title{Tab Login Modal}
\usage{
tab_login_modal(failed = FALSE)
}
\arguments{
\item{failed}{a logical value that when supplied
produces a warning message stating invalid email or
password. Default is \code{FALSE}}
}
\description{
This function produces the login modal when clicking on the
protected tabs \code{raw_data} and \code{upload_data}.
}
