% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_login.R
\name{tab_register_modal}
\alias{tab_register_modal}
\title{Tab Register Modal}
\usage{
tab_register_modal(failed = FALSE, success = FALSE)
}
\arguments{
\item{failed}{a logical value that when supplied
prodocues warning when not all fields are filled out. Default is \code{FALSE}.}

\item{success}{a logical value that when supplied
prodocues a success messsage. Default is \code{FALSE}.}
}
\description{
This function produces the registration modal on the login modal
when the registration button is slected
}
