% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxa_search_ui.R
\name{taxa_search_module}
\alias{taxa_search_module}
\alias{taxa_search_ui}
\alias{taxa_search_server}
\title{Taxa Search Tab}
\usage{
taxa_search_ui(id)

taxa_search_server(id, con)
}
\arguments{
\item{id}{the shiny namespace id name (i.e., \code{"taxa_search"}).}

\item{con}{a \code{DBI} conection to, in this case PostgreSQL database.}
}
\description{
Provides the taxa search tab.
}
\details{
\code{taxa_search_ui()} provides the taxa search user interface.

\code{taxa_search_server()} provides the taxa search server.
}
