% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_email.R
\name{taxa_submission_email_body}
\alias{taxa_submission_email_body}
\title{Taxa Submission Email Body}
\usage{
taxa_submission_email_body(n_rows, submitted_by)
}
\arguments{
\item{n_rows}{the number of rows being submitted}

\item{submitted_by}{the email of the submitter}
}
\value{
HTML of the email structure to be sent.
}
\description{
This function drafts the text body the taxa submission email that
will be sent once the submit to manager has been hit.
}
