% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_zoom.R
\name{update_zoom_slider}
\alias{update_zoom_slider}
\title{Update Zoom Slider}
\usage{
update_zoom_slider(id, vec)
}
\arguments{
\item{id}{the shiny id used e.g., \code{zoom_x}.}

\item{vec}{a \code{vector} containing the minamum and maximum values.}
}
\value{
an updated Input object for a given \verb{zoom_*} object.
}
\description{
This updates the zoom sliders
}
