% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_data.R
\name{upload_data_module}
\alias{upload_data_module}
\alias{upload_data_ui}
\alias{upload_data_server}
\title{Upload Data Tab}
\usage{
upload_data_ui(id)

upload_data_server(id, con, auth_state)
}
\arguments{
\item{id}{the shiny namespace id name (i.e., \code{"insert_data"}).}

\item{con}{a \code{DBI} conection to, in this case PostgreSQL database.}

\item{auth_state}{the authorization state to login.}
}
\description{
Provides the upload data tab.
}
\details{
\code{upload_data_ui()} provides the upload data user interface.

\code{upload_data_server()} provides the upload data server .
}
