% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_upload.R
\name{upload_to_db}
\alias{upload_to_db}
\title{Upload New Data to Database}
\usage{
upload_to_db(con, tables_to_submit)
}
\arguments{
\item{con}{a \code{DBI} conection to, in this case PostgreSQL database.}

\item{tables_to_submit}{a \code{list} containing the the tables to
be submitted.}
}
\value{
a \code{list} that has results information and the submitted
tables.
}
\description{
This function takes the split tables and uploads them to the
database.
}
