% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_vectors.R
\name{vector_functions}
\alias{vector_functions}
\alias{data_tables}
\alias{data_types}
\alias{good_groups}
\alias{not_null_fields}
\alias{numeric_fields}
\alias{protected_tabs}
\alias{sample_required_fields}
\alias{sources_required_fields}
\alias{submission_required_fields}
\alias{themes}
\title{Vector functions}
\usage{
data_tables()

data_types()

good_groups()

not_null_fields()

numeric_fields()

protected_tabs()

sample_required_fields()

sources_required_fields()

submission_required_fields()

themes()
}
\description{
These functions produce a vector with information related to
the database. These may eventually be moved to being data
objects that are loaded when the package is loaded.
}
