% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_view_data.R
\name{view_data_module}
\alias{view_data_module}
\alias{view_data_ui}
\alias{view_data_server}
\title{View Raw Data Tab}
\usage{
view_data_ui(id)

view_data_server(id, con, main_input, raw_sidebar_vals, auth_state)
}
\arguments{
\item{id}{the shiny namespace id name (i.e., \code{"view_data"}).}

\item{con}{a \code{DBI} conection to, in this case PostgreSQL database.}

\item{main_input}{the shiny input from the main server.}

\item{raw_sidebar_vals}{a \code{reactiveVal()} object produced by \code{raw_data_sidebar_server()}.
The inputs from this dictate what is displayed in the main panel.}

\item{auth_state}{the authorization state to login.}
}
\description{
Provides the raw data tab.
}
\details{
\code{view_data_ui()} provides the view data user interface.

\code{view_data_server()} provides the raw data tab server.
}
