% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_view_map.R
\name{view_map_module}
\alias{view_map_module}
\alias{view_map_ui}
\alias{view_map_server}
\title{View Map Tab}
\usage{
view_map_ui(id)

view_map_server(id, con)
}
\arguments{
\item{id}{the shiny namespace id name (i.e., \code{"view_map"}).}

\item{con}{a \code{DBI} conection to, in this case PostgreSQL database.}
}
\description{
Provides the mapping tab.
}
\details{
\code{view_map_ui()} provides the mappping user interface.

\code{view_map_server()} provides the mapping server.
}
