% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_app_version.R
\name{App version}
\alias{App version}
\alias{app_version_head}
\alias{app_version_label}
\title{App version functions}
\usage{
app_version_head()

app_version_label(app_version, build_time = NULL)
}
\arguments{
\item{app_version}{\code{character}for app version.}

\item{build_time}{a \code{POSIXct} object. If \code{NULL} it uses
\verb{Sys.time(). }}
}
\description{
These function build and display the app version
and build time.
}
