% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_check.R
\name{check_functions}
\alias{check_functions}
\alias{check_empty_character}
\alias{check_sheets}
\alias{check_tab_credentials}
\alias{check_taxonomy_match}
\title{Check functions}
\usage{
check_empty_character(x)

check_sheets(file_path, output)

check_tab_credentials(user, pass, credentials)

check_taxonomy_match(input_values, db_values)
}
\arguments{
\item{x}{a `vector}

\item{file_path}{the file path supplied by reactive handler}

\item{output}{shiny object to output too}

\item{user}{a \code{vector} containing the username}

\item{pass}{a \code{vector} containing the password}

\item{credentials}{a \code{dataframe} containing username and password}

\item{input_values}{a \code{vector} of common and species names in incoming
excel sheet.}

\item{db_values}{a \code{vector} of common and species naems that are
in \code{tbl_taxonomy}}
}
\description{
These functions check different conditions
}
