% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_clean.R
\name{clean_functions}
\alias{clean_functions}
\alias{clean_all_validations}
\alias{clean_data_tables}
\alias{clean_db_constraints}
\alias{match_to_db_col}
\alias{rename_to_db_col}
\alias{clean_row_index}
\title{Clean functions}
\usage{
clean_all_validations(...)

clean_data_tables(df, flag_cols, type, group_cols, filter_coords = TRUE)

clean_db_constraints(clause)

match_to_db_col(col_name, db_cols)

rename_to_db_col(df, con, table_name)

clean_row_index(row_index_str, min_run = 5)
}
\arguments{
\item{...}{objects to be cleaned. Usually validation reports}

\item{df}{a \code{data.frame} of imported new data}

\item{flag_cols}{collumns to flag and coalesce}

\item{type}{is the data table type that is being supplied e.g., tbl_isotope}

\item{group_cols}{columns that will be grouping our summariziing e.g.,
lat and lon, pi name, waterbody ect.}

\item{filter_coords}{Logical value that controls whether \code{NA} coords are removed.
Default value is \code{TRUE}.}

\item{clause}{constraint clause to be cleaned}

\item{col_name}{column name to be cleaned to match database column names}

\item{db_cols}{column names in the database to be used}

\item{con}{PostgreSQL connection object using \code{{DBI}}}

\item{table_name}{the table name in the database that each
column name is assinged to.}

\item{row_index_str}{a column containing the row index from submitted data during
validation}

\item{min_run}{minimum number of consecutive numbers before it condenses
the row numbers. Default is \code{5}.}
}
\description{
These functions clean up \code{data.frames} and \code{vectors} that
are coming into the the databae prior to submission,
validation reports, or are in the database and need need to be
cleaned prior to being displayed (e.g., schema constraints)
}
