% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_document_ui.R
\name{docs_module}
\alias{docs_module}
\alias{docs_ui}
\alias{docs_server}
\title{Documentation User Interface}
\usage{
docs_ui(id)

docs_server(id)
}
\arguments{
\item{id}{the shiny namespace id name (i.e., \code{"docs"}).}
}
\value{
\code{docs_ui()} a shiny UI.

\code{docs_server()} a shiny module server.
}
\description{
Provides the User Interface (UI) for the documentation tab

Provides the server for the documentation tab
}
