% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_home_tab.R
\name{home_module}
\alias{home_module}
\alias{home_tab_ui}
\title{Home Tab}
\usage{
home_tab_ui(id)
}
\arguments{
\item{id}{the shiny namespace id name (i.e., \code{"home"}).}
}
\description{
Provides the the home tab for the app.
}
\details{
\code{docs_ui()} a shiny UI.
}
