% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_how_to_use_ui.R
\name{how_to_use_module}
\alias{how_to_use_module}
\alias{how_to_ui}
\title{How to Use Tab}
\usage{
how_to_ui(id)
}
\arguments{
\item{id}{the shiny namespace id name (i.e., \code{"how_to_use"}).}
}
\description{
Provides the How to Use tab in the app.
}
\details{
\code{how_to_ui()} a shiny UI.
}
