% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_plot_control.R
\name{plot_controls}
\alias{plot_controls}
\alias{plot_ui}
\alias{pallete_selector}
\alias{alpha_selector}
\alias{size_selector}
\alias{shape_selector}
\title{Plot control functions}
\usage{
plot_ui(title, plot_id, height, ...)

pallete_selector(...)

alpha_selector(...)

size_selector(...)

shape_selector(...)
}
\arguments{
\item{title}{the title of plot}

\item{plot_id}{the shiny namespace for the plot desired \code{"scatter_plot"}}

\item{height}{the height of the plot in pixels}

\item{...}{additional parameters to be given for example a \code{NS()} object.}
}
\description{
These functions add different control features to different
plots displayed such as scatter and eventually box plot.
}
\details{
\code{plot_ui()} provides user interface for any plot object throughout
the app.

\code{pallete_selector()} provides user interface to select different palletes
from \href{https://CRAN.R-project.org/package=viridis}{viridis}.

\code{alpha_selector()} provides user interface to select different values
of \code{alpha} which controls transparency.

\code{size_selector()} provides user interface to select different values
of \code{size} which controls the size of the points.

\code{shape_selector()} provides user interface to select different values
of \code{shapes} which controls the shape type.
}
