% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_register_feilds.R
\name{register_fields}
\alias{register_fields}
\title{This function produces the UI for the registraiton fields both
in the modal and tab}
\usage{
register_fields(...)
}
\value{
shinmy UI objects
}
\description{
This function produces the UI for the registraiton fields both
in the modal and tab
}
