% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scatter_sidebar.R,
%   R/mod_view_scatter_plot.R
\name{scatter_plot_module}
\alias{scatter_plot_module}
\alias{scatter_sidebar_ui}
\alias{scatter_sidebar_server}
\alias{view_scatter_plot_ui}
\alias{scatter_plot_server}
\title{Scatter Plot Sidebar}
\usage{
scatter_sidebar_ui(id)

scatter_sidebar_server(id, con, main_input)

view_scatter_plot_ui(id)

scatter_plot_server(id, con, main_input, scatter_sidebar_vals)
}
\arguments{
\item{id}{the shiny namespace id name (i.e., \code{"scatter_plot"}).}

\item{con}{a \code{DBI} conection to, in this case PostgreSQL database.}

\item{main_input}{the shiny input from the main server.}

\item{scatter_sidebar_vals}{a \code{reactiveVal()} object produced by \code{scatter_sidebar_server()}.
The inputs from this dictate what is displayed in the main panel.}
}
\description{
Provides the sidebar for the scatter plot pane.

Provides the scatter plot tab.
}
\details{
\code{scatter_sidebar_ui()} that provides user interface for scatter plot
sidebar.

\code{scatter_sidebar_server()} provides the scatter plot sidebar server.

\code{view_scatter_plot_ui()} provides the scatter plot user interface.

\code{scatter_plot_server()} provides the scatter plot server.
}
