% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_view_source.R
\name{source_module}
\alias{source_module}
\alias{view_source_ui}
\alias{view_source_server}
\title{Source Tab}
\usage{
view_source_ui(id)

view_source_server(id, con, main_input, source_sidebar_vals)
}
\arguments{
\item{id}{the shiny namespace id name (i.e., \code{"view_source"}).}

\item{con}{a \code{DBI} conection to, in this case PostgreSQL database.}

\item{main_input}{the shiny input from the main server.}

\item{source_sidebar_vals}{a \code{reactiveVal()} object produced by \code{source_sidebar_server()}.
The inputs from this dictate what is displayed in the main panel.}
}
\description{
Provides the source tab.
}
\details{
\code{view_source_ui()} provides the source user interface.

\code{view_source_server()} provides the source server.
}
