% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_validate.R
\name{validation_functions}
\alias{validation_functions}
\alias{the_golden_lance}
\alias{validate_tbl_sources}
\alias{validate_tbl_submission}
\title{Validator funtions}
\usage{
the_golden_lance(df)

validate_tbl_sources(df)

validate_tbl_submission(df)
}
\arguments{
\item{df}{a \code{data.frame} that requires validation.}
}
\value{
returns a \code{validate} object.
}
\description{
These functions work within their own enviornment. They take
in a \code{data.frame} and a set of \code{rules} either supplied by
\verb{rule_*} functions or within the function itself to check. Once
both the \code{data.frame} and \code{rules} are set these functions use
\code{confront()} from \code{{validate}} to confront
the \code{data.frame} with rules.
}
\details{
\code{the_golden_lance()} is named after Galtar's most prized possesion, his golden lance, which
he uses to fight evil. This function validates incoming data to be uploaded to \code{tbl_samples}.
Let \code{the_golden_lance()} have its say.

\code{validate_tbl_sources()} validates incoming data to be uploaded to \code{tbl_sources}.

\code{validate_tbl_submission()} validates incoming data to be uploaded to \code{tbl_submission}.
}
