% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_data_types.R
\name{vector_functions}
\alias{vector_functions}
\alias{data_tables}
\alias{data_types}
\alias{themes}
\title{Vector functions}
\usage{
data_tables()

data_types()

themes()
}
\description{
These functions produce a vector with information related to
the database. These may eventually be moved to being data
objects that are loaded when the package is loaded.
}
